Fixed missing initialization for persistence diagram norms

+++ persistenceDiagrams/Extraction.hh
@@ -60,7 +60,7 @@ void weightedPersistence( const PersistenceDiagram<DataType>& D,
   for( auto&& point : uniquePoints )
   {
     auto count  = std::count( points.begin(), points.end(), point );
-    auto weight = count / static_cast<double>( points.size() );
+    auto weight = static_cast<double>( count ) / static_cast<double>( points.size() );
 
     *result++ = point.persistence() / weight;
   }
